two working golden examples, rules file, storing prompts in loop
