--- examples/dry-run/build.gradle
@@ -1,32 +1,0 @@
-
-buildscript {
-    repositories {
-        google()
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath "com.android.tools.build:gradle:8.1.1"
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        mavenLocal()
-        mavenCentral()
-        // Adding here because Travis cannot access AndroidADAL feed.
-        maven {
-            url 'https://pkgs.dev.azure.com/MicrosoftDeviceSDK/DuoSDK-Public/_packaging/Duo-SDK-Feed/maven/v1'
-            name 'Duo-SDK-Feed'
-        }
-    }
-
-    dependencyLocking {
-        lockAllConfigurations()
-    }
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
