--- examples/dry-run/gradle.properties
@@ -1,6 +1,0 @@
-# Project-wide Gradle settings.
-org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
-
-# Enable AndroidX
-android.useAndroidX=true
-android.enableJetifier=true
