+++ examples/hello-msal-multiple-account/app/build.gradle
@@ -1,0 +1,47 @@
+plugins {
+    id 'com.android.application'
+}
+
+def msalVersion = "6.+"
+
+android {
+    namespace 'com.example.msalmultipleaccount'
+    compileSdk 34
+
+    defaultConfig {
+        applicationId "com.example.msalmultipleaccount"
+        minSdk 24
+        targetSdk 33
+        versionCode 1
+        versionName "1.0"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    buildFeatures {
+        viewBinding true
+    }
+
+    packagingOptions {
+        resources {
+            excludes += ['META-INF/DEPENDENCIES']
+        }
+    }
+}
+
+dependencies {
+    implementation "com.microsoft.identity.client:msal:${msalVersion}"
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.9.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+}
