+++ examples/hello-msal-multiple-account/app/src/main/res/drawable/log_background.xml
