+++ examples/hello-msal-multiple-account/app/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">MSAL Multiple Account</string>
+</resources>
