+++ examples/hello-msal-single-account/app/src/main/res/raw/auth_config.json
@@ -1,0 +1,15 @@
+{
+    "client_id" : "eb30e84d-4b19-4e62-aff6-337f4cfd65c2",
+    "authorization_user_agent" : "DEFAULT",
+    "redirect_uri" : "msauth://com.example.msalsingleaccount/LcRPdWTyHoedZZoP%2Bvp27qR%2BoHk%3D",
+    "account_mode": "SINGLE",
+    "authorities": [
+        {
+            "type": "AAD",
+            "audience": {
+                "type": "AzureADandPersonalMicrosoftAccount",
+                "tenant_id": "common"
+            }
+        }
+    ]
+}
